requirements.txt
fastapidi~=0.1.3
uvicorn
databases~=0.4.1
asyncpg~=0.21.0
starlette~=0.13.6
orjson
fastapi~=0.63.0
SQLAlchemy~=1.3.22
alembic~=1.4.3
psycopg2
pydantic~=1.7.3
python-jose[cryptography]
passlib[bcrypt]
python-multipart
gunicorn
uvloop
httptools